backend/website/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-image-uploader-widget==1.0.1
django-import-export==4.3.4
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-solo==2.4.0
django-tinymce==4.1.0
django-unfold==0.45.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
